Move serializable matchers to internal testing
