+++ subprojects/core-exec/src/test/groovy/dev/nokee/core/exec/CommandLineToolInvocationEnvrionmentVariablesSerializableTests.java
@@ -18,7 +18,7 @@ package dev.nokee.core.exec;
 import com.google.common.collect.ImmutableMap;
 import org.junit.jupiter.api.Test;
 
-import static dev.nokee.core.exec.SerializableMatchers.isSerializable;
+import static dev.nokee.internal.testing.SerializableMatchers.isSerializable;
 import static org.hamcrest.MatcherAssert.assertThat;
 
 class CommandLineToolInvocationEnvrionmentVariablesSerializableTests {
